backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
beartype==0.15.0
bleach==6.3.0
bokeh==3.8.2
boto3==1.42.66
botocore==1.42.66
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
decorator==5.2.1
deltalake==0.17.4
diskcache==5.6.3
distro==1.9.0
executing==2.2.1
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fs==2.4.16
fsspec==2026.2.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.5.0
google-cloud-pubsub==2.35.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h3==4.4.2
hf-xet==1.4.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.16
linkify-it-py==2.1.0
litellm==1.82.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.7.1
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==3.0.1
panel==1.8.9
param==2.3.2
parso==0.8.6
pathway==0.29.1
pexpect==4.9.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-sat==1.9.dev2
pyviz_comms==3.0.6
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
stack-data==0.6.3
starlette==0.52.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webencodings==0.5.1
widgetsnbextension==4.0.15
xyzservices==2025.11.0
yarl==1.23.0
zipp==3.23.0
